moving lava session
